+++ components/atoms/HeaderTitle.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import { StyleSheet, Text } from  'react-native';
+
+export default class HeaderTitle extends React.Component {
+  render() {
+    return (
+      <Text style={styles.title}>{this.props.title}</Text>
+    );
+  }
+}
+
+const styles = StyleSheet.create({
+  title: {
+    fontWeight: '700',
+    fontSize: 35,
+  },
+});
